House_Price_Predictor/pyproject.toml
[project]
name = "house-price-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.1",
    "ipykernel>=7.2.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.9",
    "pandas>=3.0.3",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.49",
    "torch>=2.12.0",
    "uvicorn>=0.47.0",
    "xgboost>=3.2.0",
]
